+++ routes/scheddules.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template,redirect,url_for
+from flask import Blueprint, render_template,redirect,url_for,flash
 
 from models.schedaule import Schedule
 from extinsion import db
@@ -15,6 +15,9 @@ def show_schedules():
 @schedule.route("/admin/schedule/<int:schedule_id>/delete")
 def delete_schedule(schedule_id):
     schedule = Schedule.query.get_or_404(schedule_id)
+    if schedule.bookings:
+        flash("you can't delete this group because it has booking","danger")
+        return redirect(url_for("schedule.show_schedules"))    
     db.session.delete(schedule)
     db.session.commit()
     return redirect(url_for("schedule.show_schedules"))
